Reorganization of ros1 native tests
