+++ roslibrust/src/ros1/publisher.rs
@@ -287,75 +287,3 @@ impl From<serde_rosmsg::Error> for PublisherError {
         Self::SerializingError(value.to_string())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::ros1::NodeHandle;
-
-    // TODO stop redundantly generating the same messages!!!
-    roslibrust_codegen_macro::find_and_generate_ros_messages!("assets/ros1_common_interfaces");
-
-    // TODO should probably move all ros1 integration tests to a single file in tests/
-    #[test_log::test(tokio::test)]
-    async fn test_latching() {
-        let nh = NodeHandle::new("http://localhost:11311", "test_latching")
-            .await
-            .unwrap();
-
-        // Create a publisher that is latching
-        let publisher = nh
-            .advertise::<std_msgs::String>("/test_latching", 1, true)
-            .await
-            .unwrap();
-
-        // Publish message to no one
-        publisher
-            .publish(&std_msgs::String {
-                data: "test".to_owned(),
-            })
-            .await
-            .unwrap();
-
-        // Create a subscriber that will connect to the publisher
-        let mut subscriber = nh
-            .subscribe::<std_msgs::String>("/test_latching", 1)
-            .await
-            .unwrap();
-
-        // Try to get message from subscriber
-        let msg = subscriber.next().await.unwrap().unwrap();
-
-        // Confirm we got the message we published
-        assert_eq!(msg.data, "test");
-    }
-
-    #[test_log::test(tokio::test)]
-    async fn test_not_latching() {
-        // Opposite of test_latching, confirms no message appears when latching is false
-        let nh = NodeHandle::new("http://localhost:11311", "test_not_latching")
-            .await
-            .unwrap();
-
-        let publisher = nh
-            .advertise::<std_msgs::String>("/test_not_latching", 1, false)
-            .await
-            .unwrap();
-
-        publisher
-            .publish(&std_msgs::String {
-                data: "test".to_owned(),
-            })
-            .await
-            .unwrap();
-
-        let mut subscriber = nh
-            .subscribe::<std_msgs::String>("/test_not_latching", 1)
-            .await
-            .unwrap();
-
-        let res =
-            tokio::time::timeout(tokio::time::Duration::from_millis(250), subscriber.next()).await;
-        // Should timeout
-        assert!(res.is_err());
-    }
-}
