+++ roslibrust/src/ros1/service_client.rs
@@ -217,115 +217,3 @@ impl ServiceClientLink {
         }
     }
 }
-
-#[cfg(feature = "ros1_test")]
-#[cfg(test)]
-mod test {
-    use log::info;
-
-    use crate::ros1::{NodeError, NodeHandle};
-
-    roslibrust_codegen_macro::find_and_generate_ros_messages!(
-        "assets/ros1_test_msgs",
-        "assets/ros1_common_interfaces"
-    );
-
-    // Some logic in the service client specifically for handling large payloads
-    // trying to intentionally exercise that path
-    // TODO this could probably be moved to integration tests or own file
-    #[test_log::test(tokio::test)]
-    async fn test_large_service_payload_client() {
-        let nh = NodeHandle::new(
-            "http://localhost:11311",
-            "test_large_service_payload_client",
-        )
-        .await
-        .unwrap();
-
-        // Advertise a service that just echo's the bytes back
-        let _handle = nh
-            .advertise_service::<test_msgs::RoundTripArray, _>("large_service_payload", |request| {
-                Ok(test_msgs::RoundTripArrayResponse {
-                    bytes: request.bytes,
-                })
-            })
-            .await
-            .unwrap();
-
-        // Picking random value that should be larger than MTU
-        // Making sure the ROS message gets split over multiple TCP transactions
-        // and that we correctly re-assemble it on the other end
-        let bytes = vec![0; 10_000];
-
-        info!("Starting service call");
-        let response = nh
-            .service_client::<test_msgs::RoundTripArray>("large_service_payload")
-            .await
-            .unwrap()
-            .call(&test_msgs::RoundTripArrayRequest {
-                bytes: bytes.clone(),
-            })
-            .await
-            .unwrap();
-        info!("Service call complete");
-
-        assert_eq!(response.bytes, bytes);
-    }
-
-    #[test_log::test(tokio::test)]
-    async fn error_on_unprovided_service() {
-        let nh = NodeHandle::new("http://localhost:11311", "error_on_unprovided_service")
-            .await
-            .unwrap();
-
-        let client = nh
-            .service_client::<test_msgs::RoundTripArray>("unprovided_service")
-            .await;
-        assert!(client.is_err());
-        // Note / TODO: this currently returns an IoError(Kind(ConnectionAborted))
-        // which is better than hanging, but not a good error type to return
-        if !matches!(client, Err(NodeError::IoError(_))) {
-            panic!("Unexpected error type");
-        }
-    }
-
-    #[test_log::test(tokio::test)]
-    async fn persistent_client_can_be_called_multiple_times() {
-        let nh = NodeHandle::new(
-            "http://localhost:11311",
-            "/persistent_client_can_be_called_multiple_times",
-        )
-        .await
-        .unwrap();
-
-        let server_fn = |request: test_msgs::AddTwoIntsRequest| {
-            Ok(test_msgs::AddTwoIntsResponse {
-                sum: request.a + request.b,
-            })
-        };
-
-        let _handle = nh
-            .advertise_service::<test_msgs::AddTwoInts, _>(
-                "/persistent_client_can_be_called_multiple_times/add_two",
-                server_fn,
-            )
-            .await
-            .unwrap();
-
-        let client = nh
-            .service_client::<test_msgs::AddTwoInts>(
-                "/persistent_client_can_be_called_multiple_times/add_two",
-            )
-            .await
-            .unwrap();
-
-        for i in 0..10 {
-            let call: test_msgs::AddTwoIntsResponse = client
-                .call(&test_msgs::AddTwoIntsRequest { a: 1, b: i })
-                .await
-                .unwrap();
-
-            assert_eq!(call.sum, 1 + i);
-        }
-    }
-}
