+++ roslibrust/src/ros1/service_server.rs
@@ -239,174 +239,3 @@ impl ServiceServerLink {
         }
     }
 }
-
-#[cfg(feature = "ros1_test")]
-#[cfg(test)]
-mod test {
-    use log::*;
-    use tokio::time::timeout;
-
-    use crate::ros1::NodeHandle;
-
-    // TODO we're regenerating the messages in a lot of places
-    // makes the tests slower to compile and run
-    // we should probably generate messages for tests once in a central place?
-    roslibrust_codegen_macro::find_and_generate_ros_messages!(
-        "assets/ros1_common_interfaces",
-        "assets/ros1_test_msgs"
-    );
-
-    #[test_log::test(tokio::test)]
-    async fn basic_service_server() {
-        const TIMEOUT: std::time::Duration = std::time::Duration::from_secs(1);
-        debug!("Getting node handle");
-        let nh = NodeHandle::new("http://localhost:11311", "/basic_service_server")
-            .await
-            .unwrap();
-
-        let server_fn = |request: test_msgs::AddTwoIntsRequest| {
-            info!("Got request: {request:?}");
-            return Ok(test_msgs::AddTwoIntsResponse {
-                sum: request.a + request.b,
-            });
-        };
-
-        // Create the server
-        debug!("Creating server");
-        let _handle = nh
-            .advertise_service::<test_msgs::AddTwoInts, _>(
-                "/basic_service_server/add_two",
-                server_fn,
-            )
-            .await
-            .unwrap();
-
-        // Make the request
-        debug!("Calling service");
-        let call: test_msgs::AddTwoIntsResponse = timeout(
-            TIMEOUT,
-            timeout(
-                TIMEOUT,
-                nh.service_client::<test_msgs::AddTwoInts>("/basic_service_server/add_two"),
-            )
-            .await
-            .unwrap()
-            .unwrap()
-            .call(&test_msgs::AddTwoIntsRequest { a: 1, b: 2 }),
-        )
-        .await
-        .unwrap()
-        .unwrap();
-
-        assert_eq!(call.sum, 3);
-        debug!("Got 3");
-    }
-
-    #[test_log::test(tokio::test)]
-    async fn dropping_service_server_kill_correctly() {
-        debug!("Getting node handle");
-        let nh = NodeHandle::new("http://localhost:11311", "/dropping_service_node")
-            .await
-            .unwrap();
-
-        let server_fn = |request: test_msgs::AddTwoIntsRequest| {
-            info!("Got request: {request:?}");
-            return Ok(test_msgs::AddTwoIntsResponse {
-                sum: request.a + request.b,
-            });
-        };
-
-        // Create the server
-        let handle = nh
-            .advertise_service::<test_msgs::AddTwoInts, _>("~/add_two", server_fn)
-            .await
-            .unwrap();
-
-        // Make the request (should succeed)
-        let client = nh
-            .service_client::<test_msgs::AddTwoInts>("~/add_two")
-            .await
-            .unwrap();
-        let _call: test_msgs::AddTwoIntsResponse = client
-            .call(&test_msgs::AddTwoIntsRequest { a: 1, b: 2 })
-            .await
-            .unwrap();
-
-        // Shut down the server
-        std::mem::drop(handle);
-        // Wait a little bit for server shut down to process
-        tokio::time::sleep(std::time::Duration::from_millis(250)).await;
-
-        // Make the request again (should fail)
-        let call_2 = client
-            .call(&test_msgs::AddTwoIntsRequest { a: 1, b: 2 })
-            .await;
-        debug!("Got call_2: {call_2:?}");
-        assert!(
-            call_2.is_err(),
-            "Shouldn't be able to call after server is shut down"
-        );
-
-        // Create a new clinet
-        let client = nh
-            .service_client::<test_msgs::AddTwoInts>("~/add_two")
-            .await;
-        // Client should fail to create as there should be no provider of the service
-        assert!(
-            client.is_err(),
-            "Shouldn't be able to connect again (no provider of service)"
-        );
-
-        // Confirm ros master no longer reports our service as provided (via rosapi for fun)
-        let rosapi_client = nh
-            .service_client::<rosapi::Services>("/rosapi/services")
-            .await
-            .unwrap();
-        let service_list: rosapi::ServicesResponse = rosapi_client
-            .call(&rosapi::ServicesRequest {})
-            .await
-            .unwrap();
-        assert!(!service_list
-            .services
-            .contains(&"/dropping_service_node/add_two".to_string()));
-    }
-
-    #[test_log::test(tokio::test)]
-    async fn service_error_behavior() {
-        debug!("Getting node handle");
-        let nh = NodeHandle::new("http://localhost:11311", "/service_error_behavior")
-            .await
-            .unwrap();
-
-        let server_fn = |request: test_msgs::AddTwoIntsRequest| -> Result<
-            test_msgs::AddTwoIntsResponse,
-            Box<dyn std::error::Error + Send + Sync>,
-        > {
-            info!("Got request: {request:?}");
-            return Err(Box::new(std::io::Error::new(
-                std::io::ErrorKind::NotFound,
-                "test message",
-            )));
-        };
-
-        // Create the server
-        let _handle = nh
-            .advertise_service::<test_msgs::AddTwoInts, _>("~/add_two", server_fn)
-            .await
-            .unwrap();
-
-        // Make the request (should fail)
-        let client = nh
-            .service_client::<test_msgs::AddTwoInts>("~/add_two")
-            .await
-            .unwrap();
-        let call = client
-            .call(&test_msgs::AddTwoIntsRequest { a: 1, b: 2 })
-            .await;
-        // Okay so this is logging the error message correctly, but the contents currently suck:
-        // "Got call: Err(IoError(Custom { kind: Other, error: "Failure response from service server: Custom { kind: NotFound, error: \"test message\" }" }))"
-        // We should someday clean up error types here, but frankly errors throughout the entire crate need an overhaul
-        debug!("Got call: {call:?}");
-        assert!(call.is_err());
-    }
-}
