+++ roslibrust/src/ros1/subscriber.rs
@@ -246,57 +246,3 @@ impl From<serde_rosmsg::Error> for SubscriberError {
         Self::DeserializeError(value.to_string())
     }
 }
-
-#[cfg(test)]
-mod test {
-
-    use crate::ros1::NodeHandle;
-
-    // TODO stop redundantly doing codegen so many times in tests
-    roslibrust_codegen_macro::find_and_generate_ros_messages!(
-        "assets/ros1_test_msgs",
-        "assets/ros1_common_interfaces"
-    );
-
-    #[test_log::test(tokio::test)]
-    async fn test_large_payload_subscriber() {
-        let nh = NodeHandle::new("http://localhost:11311", "/test_large_payload_subscriber")
-            .await
-            .unwrap();
-
-        let publisher = nh
-            .advertise::<test_msgs::RoundTripArrayRequest>("/large_payload_topic", 1, false)
-            .await
-            .unwrap();
-
-        let mut subscriber = nh
-            .subscribe::<test_msgs::RoundTripArrayRequest>("/large_payload_topic", 1)
-            .await
-            .unwrap();
-
-        // Give some time for subscriber to connect to publisher
-        tokio::time::sleep(tokio::time::Duration::from_millis(100)).await;
-
-        for _i in 0..10 {
-            let bytes = vec![0; 10_000];
-            publisher
-                .publish(&test_msgs::RoundTripArrayRequest {
-                    bytes: bytes.clone(),
-                })
-                .await
-                .unwrap();
-
-            match subscriber.next().await {
-                Some(Ok(msg)) => {
-                    assert_eq!(msg.bytes, bytes);
-                }
-                Some(Err(e)) => {
-                    panic!("Got error: {e:?}");
-                }
-                None => {
-                    panic!("Got None");
-                }
-            }
-        }
-    }
-}
